refactor:move glob func in glob.py
